Add files via upload
Business Objective: Customer churn is a big problem for telecommunications companies. Indeed, their annual churn rates are usually higher than 10%. For that reason, they develop strategies to keep as many clients as possible. This is a classification project since the variable to be predicted is binary (churn or loyal customer). The goal here is to model churn probability, conditioned on the customer features.
Data Set Details: The data file telecommunications_churn.csv contains a total of 19 features for 3333 customers. Each row corresponds to a client of a telecommunications company for whom it has been collected information about the type of plan they have contracted, the minutes they have talked, or the charge they pay every month.
The data set includes the following variables:
•	account_length
•	voice_mail_plan
•	voice_mail_messages
•	day_mins
•	evening_mins
•	night_mins
•	international_mins
•	customer_service_calls
•	international_plan
•	day_calls
•	day_charge
•	evening_calls
•	evening_charge
•	night_calls
•	night_charge
•	international_calls
•	international_charge
•	total_charge
•	churn: This is the target variable. It is the one that determines whether the client is still in the company or not. (1-Churn, 0-No Churn)